+++ spec/infra/harness/main.tf
@@ -1,9 +1,5 @@
-data "template_file" "test_policy" {
-  template = file("${path.module}/resources/test-policy.json.tpl")
-
-  vars = {
-    bucket_name = var.bucket_name
-  }
+locals {
+  test_policy = templatefile("${path.module}/resources/test-policy.json.tpl", { bucket_name = var.bucket_name })
 }
 
 module "encrypted_bucket" {
@@ -15,7 +11,7 @@ module "encrypted_bucket" {
 
   mfa_delete = var.mfa_delete
 
-  source_policy_json = var.include_source_policy_json ? data.template_file.test_policy.rendered : ""
+  source_policy_json = var.include_source_policy_json ? local.test_policy : ""
 
   allow_destroy_when_objects_present = var.allow_destroy_when_objects_present