+++ terraform.tf
@@ -6,9 +6,5 @@ terraform {
       source = "hashicorp/aws"
       version = "~> 3.29"
     }
-    template = {
-      source = "hashicorp/template"
-      version = "~> 2.1"
-    }
   }
 }